src/Command.test.ts
import { Command } from './Command';

class MockCommand extends Command {
    signature = 'mockCommand {argument} {--option}';
    description = 'This is a mock command for testing';

    protected handle() {
        return Promise.resolve(0);
    }
}

describe('Command', () => {
    let command: MockCommand;

    beforeEach(() => {
        command = new MockCommand();
    });

    it('should have a signature', () => {
        expect(command.signature).toBe('mockCommand {argument} {--option}');
    });

    it('should have a description', () => {
        expect(command.description).toBe('This is a mock command for testing');
    });


    it('should handle command', async () => {
        const result = await command.run(undefined, 'value', '--option')
        expect(result).toBe(0);
    });

    // Add more tests for other methods in the Command class
    // For example, if you have a method named `mockMethod` that should return `mockValue`, you can test it like this:
    // it('should return correct value from mockMethod', () => {
    //   expect(command.mockMethod()).toBe(mockValue);
    // });
});